Worked on implementing the front end code for Build Kit screen

+++ OpenELIS/version2/OpenELIS/trunk/gwt/src/org/openelis/modules/inventoryItem/client/ComponentTab.java
@@ -126,7 +126,7 @@ public class ComponentTab extends Screen {
                         table.setCell(r, 1, data.getComponentDescription());
                         break;
                     case 2:
-                        data.setQuantity((Double)val);
+                        data.setQuantity((Integer)val);
                         break;
                 }
             }
